+++ examples/letter.rs
@@ -1,7 +1,7 @@
 extern crate sensehat_screen;
 
 #[cfg(feature = "default")]
-use sensehat_screen::{font_to_frame, FontCollection, PixelColor, Screen};
+use sensehat_screen::{font_to_frame, PixelColor, Screen, FONT_COLLECTION};
 #[cfg(feature = "default")]
 use std::thread;
 #[cfg(feature = "default")]
@@ -16,15 +16,14 @@ fn main() {
 fn main() {
     let mut screen = Screen::open("/dev/fb1").unwrap();
 
-    let fonts = FontCollection::new();
     let white_50_pct = PixelColor::WHITE.dim(0.5);
 
     let letter = "Dear Toño, I am well. Thank you. Bye. - ゆにち";
 
-    let screen_text = fonts.sanitize_str(letter).unwrap();
+    let screen_text = FONT_COLLECTION.sanitize_str(letter).unwrap();
 
     for unicode in screen_text.chars() {
-        if let Some(symbol) = fonts.get(unicode) {
+        if let Some(symbol) = FONT_COLLECTION.get(unicode) {
             let frame = font_to_frame(&symbol.byte_array(), white_50_pct);
             screen.write_frame(&frame);
         }